requirements.txt
aiohttp==1.0.5
async-timeout==2.0.0
cffi==1.11.4
chardet==3.0.4
discord.py==0.16.12
multidict==4.1.0
pycparser==2.18
PyNaCl==1.0.1
six==1.11.0
websockets==3.4
youtube-dl==2018.2.4
